+++ api/src/certification/session-management/infrastructure/repositories/complementary-certification-course-result-repository.js
@@ -1,4 +1,4 @@
-import { knex } from '../../../db/knex-database-connection.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
 
 const save = async function ({
   complementaryCertificationCourseId,
