+++ api/src/shared/domain/events/index.js
@@ -5,12 +5,12 @@ import _ from 'lodash';
 import * as eventBusBuilder from '../../../../lib/infrastructure/events/EventBusBuilder.js';
 import { EventDispatcher } from '../../../../lib/infrastructure/events/EventDispatcher.js';
 import { EventDispatcherLogger } from '../../../../lib/infrastructure/events/EventDispatcherLogger.js';
-import * as complementaryCertificationCourseResultRepository from '../../../../lib/infrastructure/repositories/complementary-certification-course-result-repository.js';
 import * as badgeAcquisitionRepository from '../../../../src/evaluation/infrastructure/repositories/badge-acquisition-repository.js';
 import * as certificationAssessmentHistoryRepository from '../../../certification/evaluation/infrastructure/repositories/certification-assessment-history-repository.js';
 import * as challengeCalibrationRepository from '../../../certification/evaluation/infrastructure/repositories/challenge-calibration-repository.js';
 import * as complementaryCertificationScoringCriteriaRepository from '../../../certification/evaluation/infrastructure/repositories/complementary-certification-scoring-criteria-repository.js';
 import * as flashAlgorithmService from '../../../certification/flash-certification/domain/services/algorithm-methods/flash.js';
+import * as complementaryCertificationCourseResultRepository from '../../../certification/session-management/infrastructure/repositories/complementary-certification-course-result-repository.js';
 import * as finalizedSessionRepository from '../../../certification/session-management/infrastructure/repositories/finalized-session-repository.js';
 import * as juryCertificationSummaryRepository from '../../../certification/session-management/infrastructure/repositories/jury-certification-summary-repository.js';
 import * as supervisorAccessRepository from '../../../certification/session-management/infrastructure/repositories/supervisor-access-repository.js';
